requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai>=1.25
websockets
requests
